requirements.txt
humanize==0.5.1
mutagen==1.43.0
natsort==6.2.0
Pillow==6.2.1
psutil==5.6.7
python-magic==0.4.15
ruamel.yaml==0.16.5
tqdm==4.43.0
